test: flytt tilstandstester til OppgaveTilstandTest
settTilstandFor-reglene testes nå som rene domentest uten database
i OppgaveTilstandTest. OppfølgingMediatorTest beholder kun E2E-testen
som verifiserer full flyt inkl. persistering.
